requirements.txt
langchain==0.0.184
openai==0.27.6
pydantic==1.10.7
faiss-cpu==1.7.3
langchain-community
python-dotenv==1.0.0
sqlite-database